+++ test/commands/get.spec.js
@@ -75,13 +75,6 @@ describe("The 'get' method", function () {
                         });
                     });
 
-                    it("gets the value correctly with array syntax and the callback being in the array", function (done) {
-                        client.GET([key, function (err, res) {
-                            helper.isString(value)(err, res);
-                            done(err);
-                        }]);
-                    });
-
                     it("should not throw on a get without callback (even if it's not useful)", function (done) {
                         client.GET(key);
                         client.on('error', function(err) {