+++ test/commands/mset.spec.js
@@ -89,7 +89,7 @@ describe("The 'mset' method", function () {
 
                         it("sets the value correctly with array syntax", function (done) {
                             client.mset([key, value2, key2, value]);
-                            client.get([key, helper.isString(value2)]);
+                            client.get(key, helper.isString(value2));
                             client.get(key2, helper.isString(value, done));
                         });
                     });